requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha>=2.0.0a0
aws-cdk.aws-ec2-alpha>=2.0.0a0
pytest>=7.0.0
pytest-cov>=4.0.0
PyYAML>=6.0.0
pydantic
aws-cdk.lambda-layer-kubectl-v32
aws-cdk.aws-eks-v2-alpha
localstack